chore: extract carousel to reduce re-render
