+++ ui/components/multichain/account-overview/account-overview-layout.tsx
@@ -1,28 +1,10 @@
-import React, { useContext, useState, useCallback, useMemo } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-
-import { removeSlide, setSelectedAccount } from '../../../store/actions';
-import { CarouselWithEmptyState } from '..';
-import {
-  getAppIsLoading,
-  getRemoteFeatureFlags,
-  hasCreatedSolanaAccount,
-} from '../../../selectors';
-import { MetaMetricsContext } from '../../../contexts/metametrics';
-import {
-  MetaMetricsEventName,
-  MetaMetricsEventCategory,
-} from '../../../../shared/constants/metametrics';
-import type { CarouselSlide } from '../../../../shared/constants/app-state';
-import { useCarouselManagement } from '../../../hooks/useCarouselManagement';
-import { CreateSolanaAccountModal } from '../create-solana-account-modal';
-import { getLastSelectedSolanaAccount } from '../../../selectors/multichain';
-import DownloadMobileAppModal from '../../app/download-mobile-modal/download-mobile-modal';
+import React from 'react';
 import { NetworkConnectionBanner } from '../../app/network-connection-banner';
 import {
   AccountOverviewTabsProps,
   AccountOverviewTabs,
 } from './account-overview-tabs';
+import { Carousel } from './carousel';
 
 export type AccountOverviewLayoutProps = AccountOverviewTabsProps & {
   children: React.ReactElement;
@@ -32,116 +14,17 @@ export const AccountOverviewLayout = ({
   children,
   ...tabsProps
 }: AccountOverviewLayoutProps) => {
-  const dispatch = useDispatch();
-  const isLoading = useSelector(getAppIsLoading);
-  const remoteFeatureFlags = useSelector(getRemoteFeatureFlags);
-  const isCarouselEnabled = Boolean(remoteFeatureFlags?.carouselBanners);
-  const trackEvent = useContext(MetaMetricsContext);
-  const [hasRendered, setHasRendered] = useState(false);
-
-  const [showCreateSolanaAccountModal, setShowCreateSolanaAccountModal] =
-    useState(false);
-  const hasSolanaAccount = useSelector(hasCreatedSolanaAccount);
-  const selectedSolanaAccount = useSelector(getLastSelectedSolanaAccount);
-
-  const [showDownloadMobileAppModal, setShowDownloadMobileAppModal] =
-    useState(false);
-
-  const { slides } = useCarouselManagement({
-    enabled: isCarouselEnabled,
-  });
-
-  const slideById = useMemo(() => {
-    const m = new Map<string, CarouselSlide>();
-    slides.forEach((s: CarouselSlide) => m.set(s.id, s));
-    return m;
-  }, [slides]);
-
-  const handleCarouselClick = (id: string) => {
-    const slide = slideById.get(id);
-    const key = slide?.variableName ?? id;
-
-    if (key === 'solana') {
-      if (hasSolanaAccount && selectedSolanaAccount) {
-        dispatch(setSelectedAccount(selectedSolanaAccount.address));
-      } else {
-        setShowCreateSolanaAccountModal(true);
-      }
-    }
-
-    if (key === 'downloadMobileApp') {
-      setShowDownloadMobileAppModal(true);
-    }
-
-    trackEvent({
-      event: MetaMetricsEventName.BannerSelect,
-      category: MetaMetricsEventCategory.Banner,
-      properties: {
-        // TODO: Fix in https://github.com/MetaMask/metamask-extension/issues/31860
-        // eslint-disable-next-line @typescript-eslint/naming-convention
-        banner_name: key,
-      },
-    });
-  };
-
-  const handleRemoveSlide = (slideId: string, isLastSlide: boolean) => {
-    if (isLastSlide) {
-      trackEvent({
-        event: MetaMetricsEventName.BannerCloseAll,
-        category: MetaMetricsEventCategory.Banner,
-      });
-    }
-
-    dispatch(removeSlide(slideId));
-  };
-
-  const handleRenderSlides = useCallback(
-    (renderedSlides: CarouselSlide[]) => {
-      if (!hasRendered) {
-        renderedSlides.forEach((slide) => {
-          trackEvent({
-            event: MetaMetricsEventName.BannerDisplay,
-            category: MetaMetricsEventCategory.Banner,
-            properties: {
-              // TODO: Fix in https://github.com/MetaMask/metamask-extension/issues/31860
-              // eslint-disable-next-line @typescript-eslint/naming-convention
-              banner_name: slide.id,
-            },
-          });
-        });
-        setHasRendered(true);
-      }
-    },
-    [hasRendered, trackEvent],
-  );
-
   return (
     <>
       <div className="account-overview__balance-wrapper flex flex-col p-4 gap-4">
         <NetworkConnectionBanner />
+
         {children}
 
-        {isCarouselEnabled && (
-          <CarouselWithEmptyState
-            slides={slides}
-            isLoading={isLoading}
-            onSlideClick={handleCarouselClick}
-            onSlideClose={handleRemoveSlide}
-            onRenderSlides={handleRenderSlides}
-          />
-        )}
+        <Carousel />
       </div>
-      <AccountOverviewTabs {...tabsProps}></AccountOverviewTabs>
-      {showCreateSolanaAccountModal && (
-        <CreateSolanaAccountModal
-          onClose={() => setShowCreateSolanaAccountModal(false)}
-        />
-      )}
-      {showDownloadMobileAppModal && (
-        <DownloadMobileAppModal
-          onClose={() => setShowDownloadMobileAppModal(false)}
-        />
-      )}
+
+      <AccountOverviewTabs {...tabsProps} />
     </>
   );
 };
